Inject secrets through the GUI
Add a GUI option to append a DSB to the open file containing any
used TLS secrets from the session. The marks the file as having
unsaved changes.

Note #19128 - we don't currently have warning about saving a file
with a DSB in a format that doesn't support it, as we do with comments.

No longer export the RSA Session ID, only CLIENT_RANDOM related
information. This has been a long standing TODO.

Fix #18400