MinGW+NSIS: Remove the Fedora hard-coded cross-build sysroot
Introduce a MINGW_SYSROOT cache variable and --sysroot Python
script option so the installer can be built in other
distributions that do not use Fedora's layout.

Add a few other DLLs and use some shell globs, tested on an
Arch Linux host.

The dependency list should be generated dynamically but we're not
there yet.