CMake: Add option to download and build Lua
This adds a FETCH_lua CMake option to download and build a static
lua library as part of Wireshark's build, using CMake's
ExternalProject.

This is useful to avoid having to add a MinGW Lua 5.2 binary package
for every distribution one might want to support for cross-compilation,
for an easy to build project like Lua that was designed specifically
for embedding.

This is opt-in and should be useful for every platform where Lua 5.2
is not packaged (and there are many).

Tested using Arch Linux with cross and non-cross builds using GCC.