Qt: Draw a border on the right side in filter edit
Draw a border on the right side in capture filter and display filter
line edit when having text. This looks better when having green/red
background.

Slightly increase margin-right to make space between the background
color and the buttons.

Fix an alignment issue when using "Left align buttons" in display filter.
This feature is missing for capture filter.