Qt: Don't look for word boundaries in FollowStreamText
On long lines without whitespace, FollowStreamText (and its
QPlainTextEdit) spends most of its time when adding text looking
for word boundaries to wrap at.

For the C Array, Hex Dump, and YAML options, we insert line
endings after fixed widths, so the default wrap behavior is fine.

In other case, we don't control whether there's line breaks or white
space (and with Raw, we ensure that there *isn't* white space or
line breaks in the middle of long lines), so it's not worth an
O(N^2) performance hit looking for word boundaries. Text with words
separated by whitespace in very long lines will word wrap worse now,
but that's acceptable in exchange for the dramatic performance
improvement.

Fix #17313.

Related to #15963 ("There should be a preference for the initial
selection, so you can default to Raw, etc. instead of always having
to wait for ASCII to finish") and #15211 ("We should probably replace
FollowStreamText's QTextEdit with our own widget.")