ber: Use unique sequence ids for each nest level
In reassemble_octet_string, fragment_add_seq_next is called
repeatedly until it halts, which means that it is only called
inside the single frame. It is also called recursively, with
the result of reassemblies at a different recursion level being
used at a higher level.

The dst_ref ID is only 16 bits; the nest_level is limited to 500.
Combine them and use that as an ID for the sequences, to avoid
having multiple reassemblies in the same frame with the same key.
Such reassemblies tend to leak, particularly when they consist
of only one fragment (so there is no tvb created for the new
fragment head.)