dflter: Fix semantics of fvalue length
Do not mix wire size, a protocol property, with fvalue
length, a property of certain types of objects (sequences).

Rename ftype_length() to ftype_wire_size(). Do not return
wire_size with fvalue_length() (use ftype_wire_size()
instead).

Make the semantic check reject taking the len() of objects
that are not arrays or lists. If the (fixed) len() of a number
is somehow useful we can add a different function for that.