9p: don't use g_strerror() to map 9P2000.L error codes to strings.
errno values are *not* platform-independent; most platforms probably use
the same numeric values for errnos that V7 UNIX returned, but, after
that, it's a bit of a free-for-all.  ENAMETOOLONG ("File name too
long"), for example, is 36 on most platforms on Linux, but is 63 on *BSD
and Darwin and 78 on Solaris.

The "L" stands for "Linux"; use a value_string that maps the errno
values used on Linux for most instruction sets to the corresponding
error message.  (On some instruction sets, Linux uses different values;
I think this was for backwards compatibility with vendor UNIXes for
those platforms.  For now, we just ignore this issue.)