glusterfs: use a value_string, not g_strerror(), to decode status values.
GlusterFS status values aren't errno values; they may mostly correspond
to errno values, but, because not only are some errno values different
on different platforms, some errno values on some Linux platforms differ
from the equivalent values on most Linux platforms, using errno values
in an over-the-wire protocol is a bad idea unless you *know for certain*
that call clients and servers will run the same operating system - and,
if that operating system is Linux, the client and server either are on
the same platform or are on platforms with compatible errno values.

Therefore, we shouldn't use g_strerror() to decode them, we should have
a value_string table based on the GF_ERROR_CODE_ values.

This also means that a status value of 0, which means "success", is
reported as such, rather thn as something such as "Undefined error 0".