SDP: Copy format specific parameters and pass them to RTP
RFC 8866 6.15 "fmtp (Format Parameters)":

  [fmtp] allows parameters that are specific to a particular format to be
  conveyed in a way that SDP does not have to understand them...
  Format-specific parameters... may be any set of parameters required to be
  conveyed by SDP and given unchanged to the media tool that will use this
  format.

For some format specific parameters, it is sufficient to dissect them
immediately when receiving the SDP. We do that currently, though that
ought to be done in a more generic fashion allowing media type
dissectors to register information instead of the SDP dissector
have to understand various media types.

Other parameters change the way that the payload is handled, and need
to be accessed by payload dissectors, or the playback codecs.

Copy the format parameters into a map and put that in the RTP dynamic
payload struct. Use that to pass those parameters to RTP payload
dissectors, so that they can change how they dissect the payload.
Use that to put the parameter map in the rtp_info struct that gets
passed to the RTP analysis and RTP player, and eventually to codecs.
The codec interface needs to be changed for the latter to work.

In particular, use this to determine whether AMR and AMR-WB are
transmitted in the octet-aligned or bandwidth-efficient mode according
to the SDP and change the dissection appropriately.

Related #17608 to (the codec interface part still needs doing.)