G729 codec: Support Annex B SID/CN frames
The bcg729 library interface supports decoding SID frames, so do
so.

rtp_audio_stream would have to be reworked a bit to allow processing
of missing frames by a decoder (which bcg729 supports, but not all
decoders do) instead of generating our own silence, but this change
reduces the number of silence samples we generate.

Fix #14343