codecs: Register and lookup case-insensitively
RFC 4855 3. Mapping to SDP Parameters

  Note that the payload format (encoding) names defined in the RTP
  Profile [4] are commonly shown in upper case.  Media subtype names
  are commonly shown in lower case.  These names are case-insensitive
  in both places.

Store and lookup codec names in an ASCII case-insensitive manner,
so that RTP Player works if the codec name is sent in SDP in a
different case.

This also has the effect of making AMR playback after using Decode
As (e.g., for a RTP stream set up heuristically without an SDP) work,
because the dissector is registered with a lower-case name of "amr"
(compare Opus, which is lower-case both in the dissector short name
and in the canonical IANA media type registration.)

Related to #17608