Change frame.time_epoch to FT_ABSOLUTE_TIME
frame.time_epoch is exactly the same field as frame.time,
just using a different representation.

It was added as FT_RELATIVE_TIME because only relative
time types had support for Unix time. Now that that is
fixed, change the field type of frame.time_epoch to
FT_ABSOLUTE_TIME.

This means the field is represented differently in ek
but that's OK, the value is the same.