SDP: Use convert_string_to_hex
Use convert_string_to_hex from epan/strutil when turning a hex
string in an SDP into a tvb.

Skip over the '=' before calling it (it does skip whitespace, as
needed for h248item), which means that we can always increment the
offset the same in decode_sdp_fmtp instead of treating certain
parameters differently.