RFC 2198: Process later codecs after an exception
If we get an nonfatal error when processing a secondary codec,
report it and continue on to the next codec. That particular codec
got an error, but the we know the length and payload type from
earlier in the header, so it shouldn't stop us from dissecting
the primary codec.

Do abort on Errors that do mean that there's no point in continuing.