Refine lua tests for clarity
Add detailed comments to `test/lua/testlib.lua` to explain its intended
use.

Add descriptive group names to several of the test files so that all
tests aren't just "other". At a minimum, create a "per-frame" test group
so the affected test scripts document which tests are run on every frame
(rather than on initial file load). There could still be further
improvement in this area.

Add expected test counts to some test files which had none. This should
make it a bit easier to tell when a test isn't running the right number
of times.

`test/lua/proto.lua` was seemingly created by copying `dissector.lua`
and modifying it, but none of the proto-specific tests that were added
make use of or need the actual dissector which was copied, nor a capture
file to run over. Remove the unneeded dissector, refactor the remainder
to still run as intended, and change caller `test/suite_wslua.py` to use
the empty pcap for this test.