COPS: Use a wmem_array
A GPtrArray inside a file scoped wmem_map that contains pointers to
file scoped objects is a recipe for leaking. Switch the GPtrArray
to a file scoped wmem_array so that it doesn't leak. The implementation
doesn't need or use any of the extra API calls that a GPtrArray gives
over a GArray.