RTP Analysis: Fix jitter for packets before the previous one
After 15013ab13626d86b6fa3559eff49a768092f1b7f, the expected
time of arrival is compared to the previous packet, in an effort
to handle clock changes better. If we're doing so, then there's
some chance that the expected time arrival moves backwards. That's
not a problem for the calculation, so long as we cast to signed
integers at some point.