USB: Print Darwin frame status only when available
Individual frame status codes can differ from overall request status.
For example when one of the two frames in isochronous IN status code is
kIOReturnNotResponding, then the request status will also be set to
kIOReturnNotResponding regardless of which frame did error.

Do not repeat the request status for all frames as it brings no value
and can actually mislead the user if frames do not share the same status
code.