T.38: Support reassembly of more than two data items in a frame
Store information about more than one reassembly in the given frame.

Don't increment additional_hdlc_data_field_counter on every pass,
but use the value from the first pass in the packet data.

Store the data field item number of the first data field in a
reassembly, to identify the fragment sequence number of when in
the first frame of the reassemble.

This produces a consistent set of fragment sequence numbers that
start at 0 and increment by 1, regardless of if there are more than
two data field fragments in a frame, or if there is data starting
a new reassembly in the same frame as a signal end ending a reassembly.

This still doesn't handle the case where more than one reassembly
begins in the same frame (as opposed to one ending and another beginning
in the same frame.) That would require changing the reassembly ID to
be a value guaranteed to increment for each reassembly, instead of
using the frame number of the first fragment.

Fix #12552. Fix #5792.