DICOM: Fix leak in export objects
g_path_get_basename() allocates a new string, which means that it is
a waste to strdup the filename before passing it to the tap and calling
that. The least confusing behavior is to to do all the memory
allocations in dcm_eo_packet, since the Export Object window will
free them when it is closed.

We can change some of the allocations from file scoped to packet
scoped memory, since the dicom_eo_t doesn't need to live any longer
than the packet. Also make the string buffers const, since they are
copied in dcm_eo_packet anyway.