epan: Decrease dependent frame hash table size
The table of dependent frames is really a set, not a hash table,
as we never use the value, only the keys / frame numbers. If you
only ever store the key itself as the value, glib optimizes by not
allocating space for the values. This optimization does not occur
if e.g., NULL is always used as the value.

Use the convenience function g_hash_table_add to ensure that we
keep it a set.