RTP: Rename the struct of per-packet info
Since the RTP info stored per packet and per conversation are
different now, rename the structs to reflect what they actually
do.

Note that some of the data does not need to be stored per-packet,
as it should remain the same for the entire conversation. It is
wasteful of memory to store larger structs in each RTP packet,
rather than retrieving both conversation and per-packet data as
necessary. The only thing that really needs to be stored per
packet is the extended sequence number and timestamp information.