RTP: SSRCs on the same RTP session have their own number space
RFC 3350 Section 5.2 notes: "An SSRC is defined to identify a single
timing and sequence number space." (See also RFC 8872 Section 3.2.2.
"Note that the RTP sequence number and RTP timestamp are scoped by
the SSRC and are thus specific per RTP stream."

Therefore, the RTP conversation data needs to store the most
recently seen sequence number and timestamp scoped by SSRC, to
avoid incorrect judgments about when the sequence number and
timestamp have undergone cycles (and thus incorrect extended
sequence number and extended timestamp calculations.)

The RTP packet data, being for a single SSRC, can continue
to store only a single value for the extended sequence number
and timestamp. This means that the conversation level data
and packet data are no longer the same struct. Rename a
few things to make it clearer when the packet data is being
references versus the conversation data.

In addition to fixing certain RTP sessions with multiple SSRCs,
this better handles cases where the same 5-tuple is reused, and
the new RTP session starts transmitting packets after a SDP offer
but before an answer, in the SIP offer/answer model. In such
cases we may still associate the RTP packets before the answer
with the previous SDP, but (if this is the first time the SSRC
is used, which it should be for that RTP sender) the extended
sequence number will in both cases be that of the first cycle,
instead of possibly being associated with a later cycle due to
a different SSRC's sequence number space.

Fix #19219.