dfilter: Change "not in" behaviour to match inequality
"A not in S" is now implemented as "A and A not_in S"
instead of "not (A in S)".

"not A in S" is implemented as "not A or A not_in S".

This is to be consistent with the way inequality has historically
worked, where "A != B" is not the same as "not A == B".

Maybe we should change both propositions to have inequality
be the same as not equality instead.

Fixes #19187.