packet-cql.c: deal with NULL and zero length map and set values
In the case where the byte length is -1 (NULL value), we need to ensure the metadata offset is progressing:
skipping the (irrelevant) offset of the type fields.
In the case where the byte length is 0, we also need to ensure the offset progresses in a similar manner.

This fixes parsing of a RESULT message which has both cases in it and now it is being parsed correctly.

Signed-off-by: Yaniv Kaul <yaniv.kaul@scylladb.com>